process/requirements.txt
requests
azure-storage-blob==12.3.2
lxml

# Dependencies required for local processing and bulk DB insertion
pandas
sqlalchemy
tqdm
mysql-connector-python
openpyxl
flask
flask-cors